Phase 4a: shared run cycles, job entrypoint, canary, and status
Handler bodies move to runtime/cycles.py so the Cloud Run Job, the
legacy Cloud Functions, and local runs execute identical code; the
function handlers are now thin shims. mn-immunization-job runs one
cycle per container execution with the trigger recorded in the ledger.
The canary cycle logs in and lists records read-only per school (no
PHI), built to run the day before the query so MIIC changes alert with
slack. The ledger gains a read side and the CLI a status command that
prints recent runs and flags any run missing a terminal event.
