+++ src/mn_immunization/runtime/main.py
@@ -7,6 +7,7 @@ from mn_immunization.runtime.cli import (
     create_parser,
     handle_bulk_query_command,
     handle_get_vaccinations_command,
+    handle_status_command,
     handle_transform_command,
     load_config,
 )
@@ -54,6 +55,8 @@ def execute_command(args, config):
         handle_bulk_query_command(args, config)
     elif args.command == "get-vaccinations":
         handle_get_vaccinations_command(args, config)
+    elif args.command == "status":
+        handle_status_command(args)
     else:
         print("Unknown command:", args.command)
         sys.exit(1)
